UI部分的更新：  增加了删除键  增加了第二页面的更新键 逻辑方面的更新:  TableView双击事件完成√  删除事件完成√ 待更新项：  更新数据页面UI待完成  然后更新数据页完成数据更新后的刷新逻辑  逻辑层代码重构

+++ src/dao/PersonDao.java
@@ -38,7 +38,7 @@ public class PersonDao {
      * @return
      * @throws SQLException
      */
-    public int deletePersonById(Long id) throws SQLException {
+    public int deletePersonById(int id) throws SQLException {
         String sql = "DELETE FROM " + table + " WHERE id=?";
         QueryRunner qr = new QueryRunner(JdbcUtil.getDataSource());
         int i = qr.update(sql, new Object[]{id});
